% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fill.R
\name{is.Fill}
\alias{is.Fill}
\alias{Fill}
\title{Create an Fill object.}
\usage{
is.Fill(x)

Fill(
  foregroundColor = "lightblue",
  backgroundColor = "lightblue",
  pattern = "SOLID_FOREGROUND"
)
}
\arguments{
\item{x}{a Fill object, as returned by \code{Fill}.}

\item{foregroundColor}{a character vector specifiying the foreground color.
Any color names as returned by \code{\link[grDevices]{colors}} can be used.
Or, a hex character, e.g. "#FF0000" for red.  For Excel 95 workbooks, only a
subset of colors is available, see the constant \code{INDEXED_COLORS_}.}

\item{backgroundColor}{a character vector specifiying the foreground color.
Any color names as returned by \code{\link[grDevices]{colors}} can be used.
Or, a hex character, e.g. "#FF0000" for red.  For Excel 95 workbooks, only a
subset of colors is available, see the constant \code{INDEXED_COLORS_}.}

\item{pattern}{a character vector specifying the fill pattern style.  Valid
values come from constant \code{FILL_STYLES_}.}
}
\value{
\code{Fill} returns a list with components from the input argument,
and a class attribute "Fill".  Fill objects are used when constructing cell
styles.

\code{is.Fill} returns \code{TRUE} if the argument is of class "Fill" and
\code{FALSE} otherwise.
}
\description{
Create an Fill object, useful when working with cell styles.
}
\examples{

  fill <-  Fill()

}
\seealso{
\code{\link{CellStyle}} for using the a \code{Fill} object.
}
\author{
Adrian Dragulescu
}
