\name{POI_constants}
\alias{POI_constants}
\alias{HALIGN_STYLES_}
\alias{VALIGN_STYLES_}
\alias{BORDER_STYLES_}
\alias{FILL_STYLES_}
\alias{CELL_STYLES_}
\alias{INDEXED_COLORS_}
\title{Constants used in the project.}
\description{
Document some Apache POI constants used in the project.
}
\usage{

HALIGN_STYLES_

VALIGN_STYLES_

BORDER_STYLES_

FILL_STYLES_

CELL_STYLES_

INDEXED_COLORS_
}
%\arguments{}
\value{
 A named vector. 
}
\author{Adrian Dragulescu}
\seealso{\code{\link{CellStyle}} for using the \code{POI_constants}.}
%\examples{}


