% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsReadExcel.R
\name{readExcelSheetsIntoList}
\alias{readExcelSheetsIntoList}
\title{将多个Excel页签读取到一个列表对象中}
\usage{
readExcelSheetsIntoList(fileName = "./data-raw/file/ptc_test.xlsx",
  sheetHeaderName = "")
}
\arguments{
\item{fileName}{文件名,如xxxx.xlsx，可包含多个页签}
}
\value{
返回列表对象
}
\description{
将多个Excel页签读取到一个列表对象中
}
\examples{
readExcelSheetsIntoList
}
