% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsReadExcel.R
\name{sheetCount}
\alias{sheetCount}
\title{获取所有页签数量}
\usage{
sheetCount(fileName = "./data-raw/file/ptc_test.xlsx")
}
\arguments{
\item{fileName}{文件名}
}
\value{
返回值，整数
}
\description{
获取所有页签数量
}
\examples{
sheetCount('aaa.xlsx')
}
