% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsReadExcel.R
\name{sheetNames}
\alias{sheetNames}
\title{获取所有Excel 页签名称}
\usage{
sheetNames(fileName = "./data-raw/file/ptc_test.xlsx")
}
\arguments{
\item{fileName}{文件名}
}
\value{
返回值
}
\description{
获取所有Excel 页签名称
}
\examples{
sheetNames('aaa.xlsx')
}
