% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellProtection.R
\name{is.CellProtection}
\alias{is.CellProtection}
\alias{CellProtection}
\title{Create a CellProtection object.}
\usage{
is.CellProtection(x)

CellProtection(locked = TRUE, hidden = FALSE)
}
\arguments{
\item{x}{A CellProtection object, as returned by \code{CellProtection}.}

\item{locked}{a logical indicating the cell is locked.}

\item{hidden}{a logical indicating the cell is hidden.}
}
\value{
\code{CellProtection} returns a list with components from the input
argument, and a class attribute "CellProtection".  CellProtection objects
are used when constructing cell styles.

\code{is.CellProtection} returns \code{TRUE} if the argument is of class
"CellProtection" and \code{FALSE} otherwise.
}
\description{
Create a CellProtection object used for cell styles.
}
\examples{


  font <-  CellProtection(locked=TRUE)

}
\seealso{
\code{\link{CellStyle}} for using the a \code{CellProtection}
object.
}
\author{
Adrian Dragulescu
}
