% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{POI_constants}
\alias{POI_constants}
\alias{HALIGN_STYLES_}
\alias{VALIGN_STYLES_}
\alias{BORDER_STYLES_}
\alias{FILL_STYLES_}
\alias{CELL_STYLES_}
\alias{INDEXED_COLORS_}
\title{Constants used in the project.}
\format{
An object of class \code{numeric} of length 7.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 14.

An object of class \code{numeric} of length 19.

An object of class \code{numeric} of length 44.

An object of class \code{numeric} of length 48.
}
\usage{
HALIGN_STYLES_

VALIGN_STYLES_

BORDER_STYLES_

FILL_STYLES_

CELL_STYLES_

INDEXED_COLORS_
}
\value{
A named vector.
}
\description{
Document some Apache POI constants used in the project.
}
\seealso{
\code{\link{CellStyle}} for using the \code{POI_constants}.
}
\author{
Adrian Dragulescu
}
\keyword{datasets}
